Fix weird object tile bug
Presumabely tmx-parser won't return tiles that are not in use by a tile layer. This means that objects in object groups that use tiles with a gid that is not referenced in a tile layer will not be parsed properly.

Fixed by iterating over each row and column available in the tileset, although this produces more sprites than necessary.